move test files out of Filters/ folder
